Imporve CSS + slight html changes
Improved CSS by adding an invisible border to the elements in the NAV bar. This means there's no movement of other elements (due to using flexbox) when a border appears on hover-transform.

Also fixed an issue with the brygga-font variable so the IM Fell English SC font appears.

And, I removed the footer from rorbua.html and kristiansund.html. I'll add it once it's done, rather than redoing it later